LayoutTests/inspector/console-uncaught-exception.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test2.js"></script>
<script src="../http/tests/inspector/console-tests2.js"></script>
<script>

function test()
{
    InspectorTest.reloadPage(function() {
        InspectorTest.expandConsoleMessages();
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    });
}


window.addEventListener("load", function(e) {
    function f() {
        throw new Error("Exception in 'load' event listener.")
    }
    f();
}, false);


setTimeout(function() {
    function bar() {
        throw new Error("Exception in setTimeout callback.");
    }
    bar();
}, 0);


function a()
{
    throw new Error("Exception in inline script.")
}

function b()
{
    a();
}

b();

</script>
</head>

<body onload="runTest()">
<p>
Tests that uncaught exceptions are logged into console.<a href="https://bugs.webkit.org/show_bug.cgi?id=47250">Bug 47250.</a>
</p>

</body>
</html>
