LayoutTests/platform/chromium/inspector/console-uncaught-exception-expected.txt

CONSOLE MESSAGE: line 35: Uncaught Error: Exception in inline script.
CONSOLE MESSAGE: line 19: Uncaught Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 27: Uncaught Error: Exception in setTimeout callback.
CONSOLE MESSAGE: line 35: Uncaught Error: Exception in inline script.
CONSOLE MESSAGE: line 19: Uncaught Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 27: Uncaught Error: Exception in setTimeout callback.
Tests that uncaught exceptions are logged into console.Bug 47250.

Page reloaded.
console-uncaught-exception.html:35Uncaught Error: Exception in inline script.aconsole-uncaught-exception.html:35bconsole-uncaught-exception.html:40(anonymous function)console-uncaught-exception.html:43
console-uncaught-exception.html:19Uncaught Error: Exception in 'load' event listener.fconsole-uncaught-exception.html:19(anonymous function)console-uncaught-exception.html:21
console-uncaught-exception.html:27Uncaught Error: Exception in setTimeout callback.barconsole-uncaught-exception.html:27(anonymous function)console-uncaught-exception.html:29

